型定義を専用フォルダに移していく
