+++ src/features/core/manifest/constatants.ts
