+++ src/features/core/manifest/index.ts
@@ -1,1 +1,2 @@
 export * from "./constatants";
+export * from "./types";
