+++ src/features/core/manifest/types.ts
@@ -1,0 +1,6 @@
+export interface RuntimePluginBundleOptions {
+  outputDirectory: string;
+  dictionaryName: string;
+  pluginSnapshotName: string;
+  description: string;
+}
